Trying to fix issue
If there is currently no average for a class, every endpoint that seeks to get that class's average will error because the length of the class average text is 0, but the code expects there to be at least 18 characters (Class Average ). Simple if statement. If this fix works, I'll make a similar fix for all of the other endpoints, right now I've only done it for averages.